Refactor tensor.hpp
