--- galois/ir/matrix.hpp
@@ -1,3 +1,0 @@
-#include "galois/ir/ir.hpp"
-
-namespace galois::ir::matrix {}  // namespace galois::ir::matrix
